solved system blowup. updated to use different grappa models, base force fields and water models.
